backend/requirements.txt
fastapi
uvicorn
openai
pdfplumber
python-dotenv
requests
numpy
beautifulsoup4
playwright
